requirements.txt
sentence-transformers
faiss-cpu
numpy
pdfplumber
python-dotenv
tqdm
tavily-python
openpyxl
pandas
